[ORGA] Move Charts components in Campaign and Ui folders
