+++ orga/app/components/campaign/charts/participants-by-day-loader.hbs
