+++ orga/app/components/campaign/charts/participants-by-stage-bar.hbs
