+++ orga/app/components/campaign/charts/participants-by-stage-loader.hbs
