+++ orga/app/components/campaign/charts/participants-by-status-loader.hbs
