+++ orga/app/styles/app.scss
@@ -18,10 +18,10 @@
 @import "globals/tables";
 @import "globals/texts";
 @import "globals/titles";
+@import "components/ui";
 @import "components/banner/communication";
 @import "components/login-form";
 @import "components/campaign";
-@import "components/charts";
 @import "components/current-organization";
 @import "components/dissociate-user-modal";
 @import "components/edit-student-number-modal";
