+++ orga/app/styles/components/campaign/charts/participants-by-day.scss
