+++ orga/app/styles/components/campaign/charts/participants-by-stage.scss
