+++ orga/app/styles/components/campaign/charts/participants-by-status.scss
