+++ orga/app/styles/components/campaign/index.scss
@@ -4,6 +4,9 @@
 @import "analysis/recommendations";
 @import "analysis/recommendation-indicator";
 @import "analysis/tube-recommendation-row";
+@import "charts/participants-by-day";
+@import "charts/participants-by-stage";
+@import "charts/participants-by-status";
 @import "header/tabs";
 @import "header/title";
 @import "header/archived-banner";
