--- orga/app/styles/components/charts/index.scss
@@ -1,6 +1,0 @@
-@import "cards";
-@import "chart";
-@import "placeholders";
-@import "participants-by-day";
-@import "participants-by-stage";
-@import "participants-by-status";
