+++ orga/app/styles/components/ui/index.scss
@@ -1,0 +1,3 @@
+@import "cards";
+@import "chart";
+@import "placeholders";
