+++ orga/app/styles/components/ui/placeholders.scss
