+++ orga/tests/integration/components/campaign/charts/participants-by-status_test.js
@@ -1,15 +1,15 @@
 import { module, test } from 'qunit';
-import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
+import setupIntlRenderingTest from '../../../../helpers/setup-intl-rendering';
 import { render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 
-module('Integration | Component | Charts::ParticipantsByStatus', function(hooks) {
+module('Integration | Component | Campaign::Charts::ParticipantsByStatus', function(hooks) {
   setupIntlRenderingTest(hooks);
 
   test('it should display status for assessment campaign', async function(assert) {
     this.participantCountByStatus = [['started', 1], ['completed', 1], ['shared', 1]];
 
-    await render(hbs`<Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{true}} />`);
+    await render(hbs`<Campaign::Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{true}} />`);
 
     assert.contains('En cours (1)');
     assert.contains('En attente d\'envoi (1)');
@@ -19,7 +19,7 @@ module('Integration | Component | Charts::ParticipantsByStatus', function(hooks)
   test('it should contains tooltips for assessment campaign', async function(assert) {
     this.participantCountByStatus = [['started', 1], ['completed', 1], ['shared', 1]];
 
-    await render(hbs`<Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{true}} />`);
+    await render(hbs`<Campaign::Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{true}} />`);
 
     assert.contains('En cours : Ces participants n’ont pas encore terminé leur parcours.');
     assert.contains('En attente d’envoi : Ces participants ont terminé leur parcours mais n’ont pas encore envoyé leurs résultats.');
@@ -29,7 +29,7 @@ module('Integration | Component | Charts::ParticipantsByStatus', function(hooks)
   test('it should display status for profile collection campaign', async function(assert) {
     this.participantCountByStatus = [['completed', 1], ['shared', 1]];
 
-    await render(hbs`<Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{false}} />`);
+    await render(hbs`<Campaign::Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{false}} />`);
 
     assert.notContains('En cours (1)');
     assert.contains('En attente d\'envoi (1)');
@@ -39,7 +39,7 @@ module('Integration | Component | Charts::ParticipantsByStatus', function(hooks)
   test('it should contains tooltips for profile collection campaign', async function(assert) {
     this.participantCountByStatus = [['completed', 1], ['shared', 1]];
 
-    await render(hbs`<Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{false}} />`);
+    await render(hbs`<Campaign::Charts::ParticipantsByStatus @participantCountByStatus={{participantCountByStatus}} @isTypeAssessment={{false}} />`);
 
     assert.contains('En attente d’envoi : Ces participants n’ont pas encore envoyé leurs profils.');
     assert.contains('Profils reçus : Ces participants ont envoyé leurs profils.');
